pyproject.toml
[project]
name = "maya-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.8"
dependencies = [
    "maya-stubs>=0.4.1",
]
